Optimisation des tests de fin de partie. Ces tests sont maintenant des méthodes abstraites de Plateau.java et sont implémentées dans les classes qui en héritent. IAMonteCarlo opérationnelle (actuellement nbSimulation = 10000) Ajout de la factory pour le GomokuFactory

+++ src/GomoKu/PlateauJeu2D/Noeud.java
@@ -2,7 +2,7 @@ package GomoKu.PlateauJeu2D;
 
 /**
  *
- * @author Antoine
+ * @author Antoine CARON
  */
 public class Noeud {
 
@@ -32,7 +32,7 @@ public class Noeud {
      * nombre de simulations effectuées
      */
     public double getMoyenne() {
-        return 100 * nbVictoire / nbSimulation;
+        return (double) 100 * (double) nbVictoire / (double) nbSimulation;
     }
 
 }
